Added refresh button,sharedPreferances(store primitive data),final

+++ app/src/main/java/com/example/weatherapp/Models/Sys.kt
@@ -6,6 +6,6 @@ data class Sys (
     val type:Int,
     val massage:Double,
     val country:String,
-    val sunrise:Int,
-    val sunset: Int
+    val sunrise:Long,
+    val sunset: Long
         ):Serializable
